finished route cleanup
