+++ server/api/auth.js
@@ -1,8 +1,8 @@
 const {User} = require('../db/models')
 
 module.exports.isAuthorized = async function(req, res, next) {
-  console.log('reqqqqqqqqq', req.params)
-  console.log('userrrrrrr', req.user)
+  // console.log('reqqqqqqqqq', req.params)
+  // console.log('userrrrrrr', req.user)
   if (req.user) {
     const user = await User.findByPk(req.user.id)
     if (user.id !== Number(req.params.userId)) {
