+++ server/api/index.js
@@ -3,7 +3,7 @@ module.exports = router
 
 router.use('/users', require('./users'))
 router.use('/products', require('./products'))
-router.use('/orders', require('./orders'))
+router.use('/guest', require('./guest'))
 
 router.use((req, res, next) => {
   const error = new Error('Not Found')
