+++ server/api/users.js
@@ -1,92 +1,9 @@
 const router = require('express').Router()
-const {User, Order, OrderProduct, Product} = require('../db/models')
+const {Order, OrderProduct, Product} = require('../db/models')
 const auth = require('./auth')
 module.exports = router
 
-// router.get('/', async (req, res, next) => {
-//   try {
-//     const users = await User.findAll({
-//       // explicitly select only the id and email fields - even though
-//       // users' passwords are encrypted, it won't help if we just
-//       // send everything to anyone who asks!
-//       attributes: ['id', 'email']
-//     })
-//     res.json(users)
-//   } catch (err) {
-//     next(err)
-//   }
-// })
-
-router.post('/guest/cart/add', (req, res, next) => {
-  try {
-    if (!req.session.cart) {
-      req.session.cart = [
-        {
-          product: req.body.product,
-          quantity: Number(req.body.qty)
-        }
-      ]
-    } else {
-      const matchingItem = req.session.cart.find(
-        orderProduct => orderProduct.product.id === req.body.product.id
-      )
-
-      if (matchingItem) {
-        matchingItem.quantity += req.body.qty
-      } else {
-        req.session.cart.push({
-          product: req.body.product,
-          quantity: Number(req.body.qty)
-        })
-      }
-    }
-    res.json(req.session.cart)
-  } catch (error) {
-    next(error)
-  }
-})
-
-router.get('/guest/cart', (req, res, next) => {
-  try {
-    if (req.session.cart) {
-      res.json(req.session.cart)
-    } else {
-      res.sendStatus(404)
-    }
-  } catch (error) {
-    next(error)
-  }
-})
-
-router.put('/guest/cart', (req, res, next) => {
-  try {
-    const matchingItem = req.session.cart.find(
-      orderProduct => orderProduct.product.id === req.body.productId
-    )
-    matchingItem.quantity = Number(req.body.qty)
-    res.json(req.session.cart)
-  } catch (error) {
-    next(error)
-  }
-})
-
-router.delete('/guest/cart/:productId', (req, res, next) => {
-  try {
-    req.session.cart = req.session.cart.filter(orderProduct => {
-      if (orderProduct.product.id === req.params.id) {
-        return false
-      } else {
-        return true
-      }
-    })
-    res.sendStatus(204)
-  } catch (error) {
-    next(error)
-  }
-})
-
 router.get('/:userId/cart', auth.isAuthorized, async (req, res, next) => {
-
   try {
     const cart = await Order.findOne({
       where: {
@@ -158,3 +75,45 @@ router.post(
     }
   }
 )
+
+router.delete('/:userId/cart/:orderId/:productId', async (req, res, next) => {
+  try {
+    await OrderProduct.destroy({
+      where: {
+        orderId: req.params.orderId,
+        productId: req.params.productId
+      }
+    })
+    res.sendStatus(204)
+  } catch (error) {
+    next(error)
+  }
+})
+
+router.put('/:userId/checkout/:orderId/', async (req, res, next) => {
+  try {
+    const order = await Order.findOne({
+      where: {
+        id: req.params.orderId,
+        status: 'open'
+      }
+    })
+    await order.update({status: 'closed'})
+    await Order.create({status: 'open', userId: req.user.id})
+    console.log('order', order)
+    const orderProduct = await OrderProduct.findOne({
+      where: {
+        orderId: order.id
+      }
+    })
+    const product = await Product.findOne({
+      where: {
+        id: orderProduct.productId
+      }
+    })
+    await product.update({stock: product.stock - Number(orderProduct.quantity)})
+    res.sendStatus(204)
+  } catch (error) {
+    next(error)
+  }
+})
